Shell History Summary
Just added a table with all of the relevant examples for storing history. Have not specified that it is "all" relevant examples. 
Concern: Is lengthening the document via. a summary really fixing the (potential) underlying issue of the docs being too long and requiring a summary in the first place?